Elasticsearch

+++ src/main/java/org/wj/letsrock/domain/article/service/TagSettingService.java
@@ -1,5 +1,6 @@
 package org.wj.letsrock.domain.article.service;
 
+import org.springframework.transaction.annotation.Transactional;
 import org.wj.letsrock.domain.article.model.request.SearchTagReq;
 import org.wj.letsrock.domain.article.model.request.TagReq;
 import org.wj.letsrock.model.vo.PageResultVo;
@@ -13,7 +14,7 @@ import org.wj.letsrock.domain.article.model.dto.TagDTO;
 public interface TagSettingService {
     void saveTag(TagReq tagReq);
 
-    void deleteTag(Integer tagId);
+    void deleteTag(Long tagId);
 
     void operateTag(Integer tagId, Integer pushStatus);
